pyproject.toml
[tool.poetry]
name = "regimetry"
version = "0.1.0"
description = "Unsupervised regime detection for financial time series using embeddings and clustering."
authors = ["kjpou1"]
license = "MIT"
readme = "README.md"
packages = [{ include = "regimetry", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "*"
pandas = "*"
scikit-learn = "^1.3"
matplotlib = "*"
seaborn = "*"
umap-learn = "*"
scipy = "*"
tqdm = "*"
torch = "^2.1"
transformers = "*"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
